u32 i32

+++ incl/misc/i128.h
@@ -10,5 +10,5 @@ ostream& operator<<(ostream& s,i128 a){
 	while(a)
 		b*=10,b+=a%10,a/=10;
 	while(b!=1)
-		cout<<signed(b%10),b/=10;
+		cout<<u32(b%10),b/=10;
 	return s;}
